refactor(rpc): extract decimal arithmetic helpers
Move decimal fixed-point parsing, evaluation, and formatting helpers out of lineairdb_rpc.cc into decimal_arith.{cc,hh}.
Register decimal_arith.cc in the server build.
